EOFpipe: avoid uninitialized variables in lei tests
EINTR can happen due to spurious wakeups and lead to
uninitialized variable warnings when comparing the result of
->do_read to a numeric value.  Thus avoid EINTR by making the
pipe non-blocking on initialization.

Furthermore, our ->do_read API is overkill and not appropriate
for only detecting EOF on a pipe since it is designed for
parsing network protocols.  Relying on level-triggering makes
more sense here, since we only want to detect EOF and don't have
to worry about event loop monopolization.