add and use a DESTROY-oriented leak detector
This leak detector writes the current Perl backtrace to a
temporary file which will be unlinked on exit.  If there's a
leak, the temporary file won't be unlinked and will contain a
backtrace to the leak source.

Relying on existing cycle detectors isn't suitable for this goal
since the cause of the leaks I'm targeting is due to a bug in
Perl v5.40.x[1] which is fixed in v5.42.0.

Modifications to existing .pm files in this commit will be
reverted in the next.

[1] https://yhbt.net/mirrors/perl.git/90595091f22dad07b381fd02ff1fb7e158fd8915/s/